# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ClawOPT backend",
    "version": "1.0.0",
    "description": "Generated from the backend route registry (npm run openapi:generate). Do not edit by hand."
  },
...
  "paths": {
...
    "/api/cron/status": {
      "get": {
        "operationId": "getApiCronStatus",
        "tags": [
          "control/cron"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/cron/status",
        "x-clawopt-order": 138
      }
    },
    "/api/cron/jobs": {
      "get": {
        "operationId": "getApiCronJobs",
        "tags": [
          "control/cron"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/cron/jobs",
        "x-clawopt-order": 139
      },
      "post": {
        "operationId": "postApiCronJobs",
        "tags": [
          "control/cron"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/cron/jobs",
        "x-clawopt-order": 140
      }
    },
    "/api/cron/jobs/{id}": {
      "put": {
        "operationId": "putApiCronJobsId",
        "tags": [
          "control/cron"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/cron/jobs/:id",
        "x-clawopt-order": 141
      },
      "delete": {
        "operationId": "deleteApiCronJobsId",
        "tags": [
          "control/cron"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/cron/jobs/:id",
        "x-clawopt-order": 145
      }
    },
    "/api/cron/jobs/{id}/enable": {
      "post": {
        "operationId": "postApiCronJobsIdEnable",
        "tags": [
          "control/cron"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/cron/jobs/:id/enable",
        "x-clawopt-order": 142
      }
    },
    "/api/cron/jobs/{id}/disable": {
      "post": {
        "operationId": "postApiCronJobsIdDisable",
        "tags": [
          "control/cron"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/cron/jobs/:id/disable",
        "x-clawopt-order": 143
      }
    },
    "/api/cron/jobs/{id}/run": {
      "post": {
        "operationId": "postApiCronJobsIdRun",
        "tags": [
          "control/cron"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/cron/jobs/:id/run",
        "x-clawopt-order": 144
      }
    },
    "/api/cron/jobs/{id}/runs": {
      "get": {
        "operationId": "getApiCronJobsIdRuns",
        "tags": [
          "control/cron"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/cron/jobs/:id/runs",
        "x-clawopt-order": 146
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "clawopt_session",
        "description": "httpOnly session cookie issued by POST /api/auth/login (web UI)."
      },
      "authTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ClawOPT-Auth-Token",
        "description": "Session token for CLI and scripts."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Same session token sent as Authorization: Bearer."
      }
    }
  }
}